Added tests for adt constructors within peak call
